pyproject.toml
[project]
name = "feedback-report"
version = "0.1.0"
description = "An asynchronus feedback report generator"
authors = [
    {name = "Shaibalini Kapuri"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "reportlab (>=4.4.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
